publish clouds node created with some changes

+++ src/cad_to_pointcloud.cpp
@@ -11,27 +11,35 @@ CADToPointCloud::CADToPointCloud(std::string pointcloud_path,
 {    
     setPCpath(pointcloud_path);
 
-    CADToMesh(cad_file); // here we get _CAD_mesh
-
-    if (big_file) MeshToPointCloud(_CAD_mesh); // here we get _CAD_cloud
-    else MeshToPointCloud(cad_file); // here we get _CAD_cloud
-
-    // MeshToPointCloud(_CAD_mesh); // here we get _CAD_cloud
-
-    pointcloud = _CAD_cloud;
+    if (CADToMesh(cad_file) == 0) // here we get _CAD_mesh
+    {
+        if (big_file) MeshToPointCloud(_CAD_mesh); // here we get _CAD_cloud
+        else 
+        {
+            if(MeshToPointCloud(cad_file)==0) // here we get _CAD_cloud
+            {
+                pointcloud = _CAD_cloud;
+            } 
+        }
+    } 
 }
 
-void CADToPointCloud::CADToMesh(std::string filename)
+int CADToPointCloud::CADToMesh(std::string filename)
 {
     pcl::PolygonMesh mesh;
     pcl::io::loadPolygonFile(_pc_path+filename,mesh);
 
-    _CAD_mesh = mesh;
+    int len = mesh.cloud.row_step * mesh.cloud.height; 
+    if (len == 0)
+        return -1;
 
+    _CAD_mesh = mesh;
+    
+    return 0;
     // visualizeMesh(_CAD_mesh);
 }
 
-void CADToPointCloud::MeshToPointCloud(std::string filename)
+int CADToPointCloud::MeshToPointCloud(std::string filename)
 {
     // TODO
     // deshacer esta forma tan cutre
@@ -42,12 +50,16 @@ void CADToPointCloud::MeshToPointCloud(std::string filename)
     system(cmd.c_str());
 
     pcl::io::loadPCDFile<pcl::PointXYZ> (outputfile, *_CAD_cloud);
+    if (_CAD_cloud->size()<=0)
+        return -1;
 
     cmd = "rm " + outputfile;
     system(cmd.c_str());
+
+    return 0;
 }
 
-void CADToPointCloud::MeshToPointCloud(pcl::PolygonMesh mesh)
+int CADToPointCloud::MeshToPointCloud(pcl::PolygonMesh mesh)
 {
     pcl::fromPCLPointCloud2(mesh.cloud, *_CAD_cloud);
     // TODO esta funcion esta muy limitada. 
@@ -55,14 +67,16 @@ void CADToPointCloud::MeshToPointCloud(pcl::PolygonMesh mesh)
     // pcl_mesh_sampling input.obj output.pcd    
 
     // visualizePointCloud(_CAD_cloud);
+    return 0;
 }
 
-void CADToPointCloud::MeshToROSPointCloud(pcl::PolygonMesh mesh)
+int CADToPointCloud::MeshToROSPointCloud(pcl::PolygonMesh mesh)
 {
     pcl_conversions::fromPCL( mesh.cloud, _CAD_cloud_msg);
 
     // Hay una función de pcl_ros muy util:
     // rosrun pcl_ros pcd_to_pointcloud input.pcd periodo _frame_id:=/world
+    return 0;
 }
 
 void CADToPointCloud::setPCpath(std::string path)
